// src/lib/types.d.ts
export type Path = Array<string | number>
// ...
export interface JSONPatchOperation {
  op: 'add' | 'remove' | 'replace' | 'copy' | 'move' | 'test'
  path: string
  from?: string
  value?: unknown
}

/**
 * @typedef {{
 *   op: 'add' | 'remove' | 'replace' | 'copy' | 'move' | 'test',
 *   path: Path,
 *   from?: Path,
 *   value?: *
 * }} PreprocessedJSONPatchOperation
 */
export interface PreprocessedJSONPatchOperation {
  op: 'add' | 'remove' | 'replace' | 'copy' | 'move' | 'test'
  path: Path
  from?: Path
  value?: unknown
}

/**
 * @typedef {JSONPatchOperation[]} JSONPatchDocument
 */
export type JSONPatchDocument = JSONPatchOperation[]
// ...
export interface MultiSelection {
  type: 'multi'
  paths: Path[]
  anchorPath: Path
  focusPath: Path
  pathsMap: Record<string, string | boolean>
}

/**
 * @typedef {{type: 'after', path: Path, anchorPath: Path, focusPath: Path}} AfterSelection
 */
export interface AfterSelection {
  type: 'after'
  path: Path
  anchorPath: Path
  focusPath: Path
}

/**
 * @typedef {{type: 'inside', path: Path, anchorPath: Path, focusPath: Path}} InsideSelection
 */
export interface InsideSelection {
  type: 'inside'
  path: Path
  anchorPath: Path
  focusPath: Path
}

/**
 * @typedef {{type: 'key', path: Path, anchorPath: Path, focusPath: Path, edit?: boolean}} KeySelection
 */
export interface KeySelection {
  type: 'key'
  path: Path
  anchorPath: Path
  focusPath: Path
  edit?: boolean
}

/**
 * @typedef {{type: 'value', path: Path, anchorPath: Path, focusPath: Path, edit?: boolean}} ValueSelection
 */
export interface ValueSelection {
  type: 'value'
  path: Path
  anchorPath: Path
  focusPath: Path
  edit?: boolean
}

/**
 * @typedef {MultiSelection | AfterSelection | InsideSelection | KeySelection | ValueSelection} Selection
 */
export type Selection =
  | MultiSelection
  | AfterSelection
  | InsideSelection
  | KeySelection
  | ValueSelection
// ...
export type ClipboardValues = Array<{ key: string; value: unknown }>